+++ riman-san-back/app/Http/Controllers/Api/UserController.php
@@ -4,26 +4,36 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Http\Resources\UserResource;
+use Gate;
 use Illuminate\Http\Request;
 use App\Models\User;
 use App\Http\Requests\StoreUserRequest;
 
 class UserController extends Controller
 {
+    function __construct()
+    {
+        $this->middleware("auth:sanctum");
+    }
 
     public function index()
     {
-        $users = User::all();
-        return UserResource::collection($users);
+        try {
+            if (Gate::allows("is-admin")) {
+                $users = User::all();
+                return UserResource::collection($users);
+            } else {
+                return response()->json(['message' => 'not allow to show users.'], 403);
+            }
+        } catch (\Throwable $th) {
+            return response()->json(['message' => 'An error occurred while showing users.'], 500);
+        }
     }
-
 
     public function store(StoreUserRequest $request)
     {
         try {
-            // Create a new user
             $user = User::create($request->all());
-            // Return the user
             return response()->json(['data' => new UserResource($user)], 200);
         } catch (\Throwable $th) {
             return response()->json(['message' => 'An error occurred while creating the user'], 500);
@@ -32,17 +42,27 @@ class UserController extends Controller
 
     public function show(string $id)
     {
-        $user = User::findOrFail($id);
-        return new UserResource($user);
+        try {
+            if (Gate::allows("is-admin")) {
+                $user = User::findOrFail($id);
+                return new UserResource($user);
+            } else {
+                return response()->json(['message' => 'not allow to show user.'], 403);
+            }
+        } catch (\Throwable $th) {
+            return response()->json(['message' => 'An error occurred while showing user.'], 500);
+        }
     }
-
-
     public function update(Request $request, string $id)
     {
         try {
-            $user = User::findOrFail($id);
-            $user->update($request->all());
-            return response()->json(['data' => new UserResource($user)], 200);
+            if (Gate::allows("is-admin")) {
+                $user = User::findOrFail($id);
+                $user->update($request->all());
+                return response()->json(['data' => new UserResource($user)], 200);
+            } else {
+                return response()->json(['message' => 'not allow to update user.'], 403);
+            }
         } catch (\Throwable $th) {
             return response()->json(['message' => 'An error occurred while updating the user'], 500);
 
@@ -52,9 +72,13 @@ class UserController extends Controller
     public function destroy(string $id)
     {
         try {
-            $user = User::findOrFail($id);
-            $user->delete();
-            return response()->json(['data' => 'user deleted successfully'], 200);
+            if (Gate::allows("is-admin")) {
+                $user = User::findOrFail($id);
+                $user->delete();
+                return response()->json(['data' => 'user deleted successfully'], 200);
+            } else {
+                return response()->json(['message' => 'not allow to delete user.'], 403);
+            }
         } catch (\Throwable $th) {
             return response()->json(['message' => 'An error occurred while deleting the user'], 500);
         }
